requirements.txt
attrs==21.4.0
boto3==1.21.27
botocore==1.24.27
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.1.0
cycler==0.11.0
dice-ml==0.7.2
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.3
fonttools==4.31.2
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
huggingface-hub==0.4.0
idna==3.3
imageio==2.16.1
inflect==5.4.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.4.0
lark==1.1.2
lime==0.2.0.1
llvmlite==0.39.0
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.7.1
nltk==3.7
numba==0.56.0
numpy==1.21.6
openai==0.16.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandas-stubs==1.2.0.53
pathtools==0.1.2
patsy==0.5.2
Pillow==9.0.1
pluggy==1.0.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
s3transfer==0.5.2
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.95
sentry-sdk==1.9.5
setproctitle==1.3.2
shap==0.40.0
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
tifffile==2022.3.25
tokenizers==0.11.6
tomli==2.0.1

tqdm==4.63.1
transformers==4.17.0
twilio==7.8.0
typing_extensions==4.1.1
urllib3==1.26.12
wandb==0.13.2
Werkzeug==2.0.3
word2number==1.1
wordninja==2.0.0
